# swagger.yaml
openapi: 3.0.0
info:
  title: Auth-Service API
  version: 1.0.0
  description: A concise and robust authentication service for applications requiring user management and secure JWT-based token generation. It provides endpoints for user registration, login, token refreshing, and profile management, all secured with ECDSA-signed tokens.
servers:
  - url: /
    description: Local server
paths:
...
  /v1/login:
    post:
      summary: Login User
      operationId: LoginUser
      description: Authenticates a user with their identifier (username or email) and password, issuing JWT access and refresh tokens upon success.
      tags:
        - Authentication
      requestBody:
        $ref: '#/components/requestBodies/LoginUserBody'
      responses:
        '200':
          description: Authentication successful. Access and refresh tokens are returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Bad Request - Invalid input data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidBody:
                  value:
                    error: Invalid request body
        '401':
          description: Unauthorized - Invalid credentials provided.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidCredentials:
                  value:
                    error: Invalid credentials
        '500':
          description: Internal server error during the login process.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                tokenGenerationFailed:
                  value:
                    error: Failed to generate tokens
...
components:
  schemas:
...
    LoginUser:
      type: object
      required:
        - identifier
        - password
      properties:
        identifier:
          type: string
          description: User's username or email address.
          example: newuser
        password:
          type: string
          description: User's password.
          example: StrongPassword123
...
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
          description: A short-lived JWT for accessing protected resources.
          example: eyJhbGciOiJFUzI1Ni...
        refresh_token:
          type: string
          description: A long-lived JWT for obtaining new access tokens.
          example: eyJhbGciOiJFUzI1Ni...
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: A brief description of the error.
        details:
          type: string
          description: Optional field for more detailed error information.
  requestBodies:
...
    LoginUserBody:
      description: Credentials for user login.
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/LoginUser'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: A JWT access token is required to access this endpoint. It should be provided in the Authorization header as a Bearer token.